🚚 Organise eudemo maintenance tests
